将profit的值由Long改为Double，更准确

+++ src/main/java/ipc/Clustering.java
@@ -33,7 +33,7 @@ public interface Clustering extends VersionedProtocol {
 
 	public double getProfit();
 	
-	abstract public long[] getSizeOfNotEmptyAndProfit(double r );
+	abstract public double[] getSizeOfNotEmptyAndProfit(double r );
 	
 	abstract public int getSizeOfCluster(int clusterId);
